requirements.txt
altgraph==0.17.4
appdirs==1.4.4
asgiref==3.8.1
babel==2.17.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
bottle==0.13.2
bs4==0.0.2
build==1.2.2.post1
buildozer==1.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptografy==3.0.0
cryptography==43.0.3
customtkinter==5.2.2
cycler==0.12.1
Cython==3.0.12
darkdetect==0.8.0
distlib==0.3.9
dj-database-url==3.0.1
Django==5.1.4
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docopt==0.6.2
docutils==0.21.2
drf-yasg==1.21.10
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.57.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
ImageHash==4.3.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
Kivy==2.3.1
Kivy-Garden==0.1.5
kivymd==1.2.0
kiwisolver==1.4.8
macholib==1.16.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pexpect==4.9.0
pillow==11.1.0
pipreqs==0.4.13
platformdirs==4.3.7
pooch==1.8.2
proxy_tools==0.1.0
psycopg==3.2.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.10.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-engineio==4.12.1
python-socketio==5.13.0
pytz==2025.2
pyvista==0.45.0
PyWavelets==1.8.0
pywebview==5.4
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
scooby==0.10.0
setuptools==75.6.0
sh==1.14.3
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.3
tkcalendar==1.6.1
toml==0.10.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.2.3
virtualenv==20.29.3
vtk==9.4.2
Werkzeug==3.1.3
wsproto==1.2.0
yarg==0.1.10
